Organize LArCV input data structures into larcv/ subdirectory
- Created src/spine/data/larcv/ subdirectory for LArCV-format input data
- Moved 7 modules: crt.py, meta.py, neutrino.py, optical.py, particle.py, run_info.py, trigger.py
- Updated all relative imports in moved files (. -> ..)
- Created larcv/__init__.py with re-exports for backward compatibility
- Updated imports in spine/data/__init__.py to use larcv submodule
- Updated imports in out/ modules to use spine.data.larcv.*
- Updated test imports to use top-level spine.data imports
- Maintains full backward compatibility: from spine.data import Particle still works
- Clearer organization: larcv/ = detector input, out/ = analysis output, batch/ = utilities
