+++ test/test_data/test_meta.py
@@ -2,7 +2,7 @@
 
 import numpy as np
 
-from spine.data.meta import Meta
+from spine.data import Meta
 
 
 class TestMetaCreation:
